split out DemodulatorController from mainwindow* files
